pyproject.toml
[tool.poetry]
name = "hfsm"
version = "0.1.0"
description = ""
authors = ["Oscar Laird <olaird25@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flaml = {extras = ["automl"], version = "^2.1.1"}
sagemaker = "^2.197.0"
transformers = "^4.35.2"
datasets = {extras = ["s3"], version = "^2.15.0"}
torch = "^2.1.1"
ipywidgets = "^8.1.1"
aiohttp = "^3.9.0"
asyncio = "^3.4.3"
backoff = "^2.2.1"
nest-asyncio = "^1.5.8"
bs4 = "^0.0.1"
openai = "^1.3.5"
python-dotenv = "^1.0.0"
sentence-transformers = "^2.2.2"
faiss-cpu = "^1.7.4"
evadb = "^0.3.9"
# evadb = { path = "/home/oscar/evadb",  develop = true }
flask = "^3.0.0"
flask-cors = "^4.0.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
